+++ CHANGELOG.md
@@ -9,6 +9,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [1.3.2] - 2019-08-29
+
+### Fixed
+
+- Cleaned up some npm security issues
+
+## [1.3.1] - 2019-08-29
+
+### Fixed
+
+- Fixed an issue that prevented the whispering of themes if there were too many themes to send. This change will break the list of themes into smaller messages and send multiple whispers.
+
+## [1.3.0] - 2019-08-13
+
+### Added
+
+- `!theme install {theme unique identifier}` command that prompts the user to install or preview a theme supplied by chat.
+
+## [1.2.0] - 2019-07-30
+
+### Added
+
+- `!theme random light` command that chooses a random theme from the light theme extensions.
+- `!theme random dark` command that chooses a random theme from the dark theme extensions.
+
 ## [1.1.0] - 2019-06-26
 
 ### Added
@@ -120,7 +145,11 @@ them know it's not a valid theme name
 
 - Pre-release version to gather feedback from the community and help identify gaps.
 
-[Unreleased]: https://github.com/michaeljolley/vscode-twitch-themer/compare/1.1.0...HEAD
+[Unreleased]: https://github.com/michaeljolley/vscode-twitch-themer/compare/1.3.2...HEAD
+[1.3.1]: https://github.com/michaeljolley/vscode-twitch-themer/compare/1.3.1...1.3.2
+[1.3.1]: https://github.com/michaeljolley/vscode-twitch-themer/compare/1.3.0...1.3.1
+[1.3.0]: https://github.com/michaeljolley/vscode-twitch-themer/compare/1.2.0...1.3.0
+[1.2.0]: https://github.com/michaeljolley/vscode-twitch-themer/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/michaeljolley/vscode-twitch-themer/compare/1.0.1...1.1.0
 [1.0.1]: https://github.com/michaeljolley/vscode-twitch-themer/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/michaeljolley/vscode-twitch-themer/compare/0.2.0...1.0.0